Update pluginized sample to use PluginExtension base protocol
- Updated pluginized sample folder names to correctly reflect the core, non-core structure.
- Added base generator unit test.
