+++ Sample/MVC/Cartfile.resolved
@@ -1,3 +1,3 @@
 github "ReactiveX/RxSwift" "4.2.0"
 github "SnapKit/SnapKit" "4.0.0"
-github "uber/needle" "v0.5.2"
+github "uber/needle" "v0.7.0"
