+++ Sample/Pluginized/Cartfile
@@ -1,3 +1,3 @@
-github "https://github.com/uber/needle.git" ~> 0.6
+github "https://github.com/uber/needle.git" ~> 0.7
 github "https://github.com/ReactiveX/RxSwift.git" ~> 4.0
 github "https://github.com/SnapKit/SnapKit.git" ~> 4.0
