+++ Sample/Pluginized/TicTacToe/TicTacToeCore/AppDelegate.swift
