+++ Sample/Pluginized/TicTacToe/TicTacToeCore/Assets.xcassets/AppIcon.appiconset/Contents.json
