+++ Sample/Pluginized/TicTacToe/TicTacToeCore/Info.plist
