+++ Sample/Pluginized/TicTacToe/TicTacToeCore/LoggedIn/LoggedInComponent.swift
@@ -19,24 +19,26 @@ import ScoreSheet
 import TicTacToeIntegrations
 import UIKit
 
-protocol GameDependency: Dependency {
-    var mutableScoreStream: MutableScoreStream { get }
-    var playersStream: PlayersStream { get }
-}
-
-protocol GamePluginExtension {
+protocol LoggedInPluginExtension: PluginExtension {
     var scoreSheetBuilder: ScoreSheetBuilder { get }
+    var mutableScoreStream: MutableScoreStream { get }
 }
 
-class GameComponent: PluginizedComponent<GameDependency, GamePluginExtension, GameNonCoreComponent>, GameBuilder {
+class LoggedInComponent: PluginizedComponent<EmptyDependency, LoggedInPluginExtension, LoggedInNonCoreComponent>, LoggedInBuilder {
 
-    var gameViewController: UIViewController {
-        return GameViewController(mutableScoreStream: dependency.mutableScoreStream, playersStream: dependency.playersStream, scoreSheetBuilder: pluginExtension.scoreSheetBuilder)
+    var loggedInViewController: UIViewController {
+        return LoggedInViewController(gameBuilder: gameComponent,
+                                      scoreStream: pluginExtension.mutableScoreStream,
+                                      scoreSheetBuilder: pluginExtension.scoreSheetBuilder)
     }
+
+    var gameComponent: GameComponent {
+        return GameComponent(parent: self)
+	}
 }
 
 // Use a builder protocol to allow mocking for unit tests. At the same time,
-// this allows GameViewController to be initialized lazily.
-protocol GameBuilder {
-    var gameViewController: UIViewController { get }
+// this allows LoggedInViewController to be initialized lazily.
+protocol LoggedInBuilder {
+    var loggedInViewController: UIViewController { get }
 }
