+++ Sample/Pluginized/TicTacToe/TicTacToeCoreTests/Info.plist
